+++ apps/api/v2/src/modules/auth/guard/organization-roles/organization-roles.guard.ts
@@ -1,5 +1,5 @@
 import { Roles } from "@/modules/auth/decorator/roles/roles.decorator";
-import { MembershipRepository } from "@/modules/repositories/membership/membership-repository.service";
+import { MembershipRepository } from "@/modules/membership/membership-repository.service";
 import { Injectable, CanActivate, ExecutionContext } from "@nestjs/common";
 import { Reflector } from "@nestjs/core";
