+++ apps/api/v2/src/modules/auth/strategy/next-auth/next-auth.strategy.ts
@@ -1,4 +1,4 @@
-import { UserRepository } from "@/modules/repositories/user/user-repository.service";
+import { UserRepository } from "@/modules/user/user-repository.service";
 import { Injectable, UnauthorizedException } from "@nestjs/common";
 import { ConfigService } from "@nestjs/config";
 import { PassportStrategy } from "@nestjs/passport";
