+++ apps/api/v2/src/modules/booking/booking.controller.ts
@@ -1,8 +1,8 @@
 import { GetUser } from "@/modules/auth/decorator";
+import { BookingRepository } from "@/modules/booking/booking-repository.service";
 import { CreateBookingInput } from "@/modules/booking/input/create-booking";
 import { PrismaReadService } from "@/modules/prisma/prisma-read.service";
 import { PrismaWriteService } from "@/modules/prisma/prisma-write.service";
-import { BookingRepository } from "@/modules/repositories/booking/booking-repository.service";
 import {
   Body,
   Controller,
