+++ apps/api/v2/src/modules/booking/booking.module.ts
@@ -1,10 +1,11 @@
+import { BookingRepository } from "@/modules/booking/booking-repository.service";
 import { BookingController } from "@/modules/booking/booking.controller";
 import { PrismaModule } from "@/modules/prisma/prisma.module";
-import { BookingRepositoryModule } from "@/modules/repositories/booking/booking-repository.module";
 import { Module } from "@nestjs/common";
 
 @Module({
-  imports: [PrismaModule, BookingRepositoryModule],
+  imports: [PrismaModule],
+  providers: [BookingRepository],
   controllers: [BookingController],
 })
 export class BookingModule {}
