+++ apps/api/v2/src/modules/user/user.module.ts
@@ -1,5 +1,5 @@
 import { PrismaModule } from "@/modules/prisma/prisma.module";
-import { UserRepository } from "@/modules/repositories/user/user-repository.service";
+import { UserRepository } from "@/modules/user/user-repository.service";
 import { Module } from "@nestjs/common";
 
 @Module({
