+++ apps/api/v2/test/mocks/next-auth-mock.strategy.ts
@@ -1,4 +1,4 @@
-import { UserRepository } from "@/modules/repositories/user/user-repository.service";
+import { UserRepository } from "@/modules/user/user-repository.service";
 import { Injectable } from "@nestjs/common";
 import { PassportStrategy } from "@nestjs/passport";
