+++ apps/api/v2/test/utils/withNextAuth.ts
@@ -1,5 +1,5 @@
 import { NextAuthStrategy } from "@/modules/auth/strategy";
-import { UserRepository } from "@/modules/repositories/user/user-repository.service";
+import { UserRepository } from "@/modules/user/user-repository.service";
 import { TestingModuleBuilder } from "@nestjs/testing";
 
 import { NextAuthMockStrategy } from "../mocks/next-auth-mock.strategy";
